backend/requirements.txt
python-dotenv
mysql-connector-python
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-multipart
PyJWT
